// src/entities/user/model/types.ts
export interface IFriend {
	id: number;
	username: string;
	firstName: string;
	lastName: string;
	avatar?: string;
	status: 'accepted' | 'pending' | 'rejected';
	createdAt: string;
}